pyproject.toml
[project]
name = "oss-ai-stack-map"
version = "0.1.0"
description = "Research pipeline for mapping major open source AI repository tech stacks on GitHub"
readme = "README.md"
authors = [
    { name = "nibzard", email = "wave@nibzard.com" }
]
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.5.1",
    "httpx>=0.28.1",
    "pyarrow>=23.0.1",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "rich>=14.1.0",
    "tenacity>=9.1.2",
    "typer>=0.24.1",
]

[project.scripts]
oss-ai-stack-map = "oss_ai_stack_map:main"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
